[FLEX-338] Add app-version labels so SRS notifier can send to SRS
This change adds `app-version` labels to the DECKS and dks-testapp
application charts. This is done so that when these applications are
used as an involved object for reporting an KAHM event, the SRS notifier
will be able to read these labels and report the version in the
SRS' `ucodeVersion` field.

The value used for these labels is {{.Chart.Name}}-{{.Chart.Version}}.

For DECKS, a `DECKS_CHART_VERSION` environment variable is also added
(set to {{.Chart.Name}}-{{.Chart.Version}}). This is used by the SRS
GW CR controller to populate the `ucodeVersion` field in the test call
home connect home message.